Add role, user and CRUD Generator

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\HomeController;
+use App\Http\Controllers\MemberController;
 use App\Http\Controllers\RoleController;
 use App\Http\Controllers\SettingController;
 use App\Http\Controllers\UserController;
@@ -14,11 +15,14 @@ Route::middleware('auth')->group(function () {
     Route::prefix('role')->group(function () {
         Route::name('role.')->group(function () {
             Route::controller(RoleController::class)->group(function(){
-                Route::get('/', 'index')->name('index');
-                Route::post('/', 'create')->name('index');
-                Route::post('/{id}/edit', 'edit')->name('edit');
-                Route::put('/{id}/edit', 'update')->name('update');
-                Route::delete('/{id}/delete', 'destroy')->name('delete');
+                Route::get('/', 'index')->name('index')->middleware('role_or_permission:superadmin|view role');
+                Route::post('/', 'create')->name('create')->middleware('role_or_permission:superadmin|create role');
+                Route::post('/{id}/edit', 'edit')->name('edit')->middleware('role_or_permission:superadmin|edit role');
+                Route::put('/{id}/edit', 'update')->name('update')->middleware('role_or_permission:superadmin|edit role');
+                Route::delete('/{id}/delete', 'destroy')->name('delete')->middleware('role_or_permission:superadmin|delete role');
+                
+                Route::get('/{id}/permission', 'permission')->name('permission')->middleware('role_or_permission:superadmin|view role');
+                Route::put('/{id}/permission', 'updatePermission')->name('permission.update')->middleware('role_or_permission:superadmin|edit role');
             });
         });
     });
@@ -27,7 +31,11 @@ Route::middleware('auth')->group(function () {
     Route::prefix('user')->group(function () {
         Route::name('user.')->group(function () {
             Route::controller(UserController::class)->group(function(){
-                Route::get('/', 'index')->name('index');
+                Route::get('/', 'index')->name('index')->middleware('role_or_permission:superadmin|view user');
+                Route::post('/', 'create')->name('create')->middleware('role_or_permission:superadmin|create user');
+                Route::post('/{id}/edit', 'edit')->name('edit')->middleware('role_or_permission:superadmin|edit user');
+                Route::put('/{id}/edit', 'update')->name('update')->middleware('role_or_permission:superadmin|edit user');
+                Route::delete('/{id}/delete', 'destroy')->name('delete')->middleware('role_or_permission:superadmin|delete user');
             });
         });
     });
